refactor: Vo 객체 Vo 패키지로 이동
