+++ src/main/java/lotto/controller/LottoController.java
@@ -5,11 +5,11 @@ import lotto.controller.dto.PurchaseDto;
 import lotto.controller.dto.WinningStatisticsDto;
 import lotto.domain.LottoGenerator;
 import lotto.domain.LottoMachine;
-import lotto.domain.WinningNumbers;
 import lotto.domain.vo.Lotto;
 import lotto.domain.vo.LottoNumber;
 import lotto.domain.vo.Lottos;
 import lotto.domain.vo.Money;
+import lotto.domain.vo.WinningNumbers;
 import lotto.domain.vo.WinningResult;
 import lotto.view.ConsoleInputView;
 import lotto.view.ConsoleOutputView;
